+++ arquillian/adapter/src/main/java/org/wildfly/swarm/arquillian/ArtifactDependencies.java
