+++ arquillian/adapter/src/main/java/org/wildfly/swarm/arquillian/adapter/MainSpecifier.java
