+++ arquillian/adapter/src/main/java/org/wildfly/swarm/arquillian/adapter/SimpleContainer.java
