+++ arquillian/adapter/src/main/java/org/wildfly/swarm/arquillian/adapter/WildFlySwarmCommandService.java
