+++ arquillian/adapter/src/main/resources/arquillian.xml
