+++ arquillian/adapter/src/test/java/org/wildfly/swarm/arquillian/adapter/MavenProfileLoaderTest.java
