--- arquillian/api/src/main/java/org/wildfly/swarm/arquillian/adapter/InVM.java
@@ -1,28 +1,0 @@
-/**
- * Copyright 2015-2016 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wildfly.swarm.arquillian.adapter;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-/**
- * Signifies that an Arquillian-based test class should be run InVM instead of in a remote uberjar.
- *
- * @author Toby Crawley
- */
-@Retention(RetentionPolicy.RUNTIME)
-public @interface InVM {
-}
