+++ arquillian/arquillian/module.conf
@@ -1,0 +1,6 @@
+org.wildfly.swarm.arquillian.daemon
+org.wildfly.swarm.arquillian.adapter
+org.jboss.arquillian export=true
+org.wildfly.swarm.msc
+org.jboss.msc
+org.jboss.as.server
