+++ arquillian/arquillian/src/main/resources/modules/org/jboss/arquillian/main/module.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.3" name="org.jboss.arquillian" slot="main">
+
+  <resources>
+    <artifact name="org.jboss.arquillian.test:arquillian-test-spi:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.test:arquillian-test-api:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.test:arquillian-test-impl-base:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.container:arquillian-container-spi:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.container:arquillian-container-test-api:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.container:arquillian-container-test-spi:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.container:arquillian-container-test-impl-base:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.core:arquillian-core-api:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.core:arquillian-core-spi:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.test:arquillian-test-api:${version.org.arquillian}"/>
+    <artifact name="org.jboss.arquillian.test:arquillian-test-spi:${version.org.arquillian}"/>
+  </resources>
+
+  <dependencies>
+    <module name="javax.annotation.api" export="true"/>
+    <module name="javax.enterprise.api" export="true"/>
+    <module name="org.jboss.shrinkwrap" export="true"/>
+    <module name="javax.api"/>
+  </dependencies>
+</module>
