+++ arquillian/arquillian/src/main/resources/modules/org/wildfly/swarm/arquillian/adapter/main/module.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.arquillian.adapter" slot="main">
+
+  <resources>
+    <artifact name="org.wildfly.swarm:arquillian-adapter:${project.version}"/>
+  </resources>
+
+  <dependencies>
+    <module name="org.jboss.arquillian"/>
+    <module name="org.jboss.msc"/>
+    <module name="org.jboss.modules" export="true"/>
+  </dependencies>
+</module>
