+++ arquillian/arquillian/src/main/resources/modules/org/wildfly/swarm/arquillian/daemon/main/module.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.arquillian.daemon" slot="main">
+
+  <resources>
+    <artifact name="org.wildfly.swarm:arquillian-daemon:${project.version}"/>
+  </resources>
+
+  <dependencies>
+    <module name="io.netty"/>
+    <module name="javax.annotation.api" export="true"/>
+    <module name="javax.enterprise.api" export="true"/>
+    <module name="org.jboss.as.server"/>
+    <module name="org.jboss.modules"/>
+    <module name="org.jboss.msc"/>
+    <module name="org.jboss.shrinkwrap" export="true"/>
+    <module name="org.jboss.arquillian"/>
+  </dependencies>
+</module>
