--- arquillian/daemon/src/main/resources/modules/org/wildfly/swarm/arquillian/daemon/api/module.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.arquillian.daemon" slot="api">
-
-  <resources>
-    <artifact name="org.wildfly.swarm:arquillian-daemon:${project.version}"/>
-    <artifact name="org.jboss.arquillian.container:arquillian-container-spi:${version.org.arquillian}"/>
-    <artifact name="org.jboss.arquillian.container:arquillian-container-test-spi:${version.org.arquillian}"/>
-    <artifact name="org.jboss.arquillian.core:arquillian-core-api:${version.org.arquillian}"/>
-    <artifact name="org.jboss.arquillian.core:arquillian-core-spi:${version.org.arquillian}"/>
-    <artifact name="org.jboss.arquillian.test:arquillian-test-api:${version.org.arquillian}"/>
-    <artifact name="org.jboss.arquillian.test:arquillian-test-spi:${version.org.arquillian}"/>
-  </resources>
-
-  <dependencies>
-    <module name="io.netty"/>
-    <module name="javax.annotation.api" export="true"/>
-    <module name="javax.enterprise.api" export="true"/>
-    <module name="org.jboss.as.server"/>
-    <module name="org.jboss.modules"/>
-    <module name="org.jboss.msc"/>
-    <module name="org.jboss.shrinkwrap" export="true"/>
-    <module name="org.wildfly.swarm.container"/>
-  </dependencies>
-</module>
