--- arquillian/daemon/src/main/resources/modules/org/wildfly/swarm/arquillian/daemon/main/module.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.arquillian.daemon">
-  <dependencies>
-    <module name="org.wildfly.swarm.arquillian.daemon" slot="api" export="true" services="export"/>
-  </dependencies>
-</module>
