--- arquillian/daemon/src/main/resources/modules/org/wildfly/swarm/arquillian/daemon/runtime/module.xml
@@ -1,7 +1,0 @@
-<module xmlns="urn:jboss:module:1.3" name="org.wildfly.swarm.arquillian.daemon" slot="runtime">
-
-  <dependencies>
-    <module name="org.wildfly.swarm.container"/>
-    <module name="org.wildfly.swarm.container" slot="runtime"/>
-  </dependencies>
-</module>
