+++ pom.xml
@@ -397,8 +397,8 @@
       </dependency>
       <dependency>
         <groupId>org.wildfly.swarm</groupId>
-        <artifactId>arquillian-daemon-modules</artifactId>
-        <version>2016.9-SNAPSHOT</version>
+        <artifactId>arquillian-adapter</artifactId>
+        <version>2016.11.0-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>org.wildfly.swarm</groupId>
